pyproject.toml
[project]
name = "review-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Lauiee",email = "132264578+Lauiee@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.4,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "cryptography (>=44.0.0,<45.0.0)",
    "django-allauth (>=65.3.1,<66.0.0)",
    "openai (>=1.59.6,<2.0.0)",
    "drf-yasg (>=1.21.8,<2.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "rabbitmq (>=0.2.0,<0.3.0)",
    "django-cors-headers (>=4.6.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"